Refactor simulation export panel to more generic CSVExportPanel
